+++ src/index.html
@@ -38,10 +38,31 @@
         </select>
 
         <select name="Clasify" id="Clasify">
-          <option value="clasify">Clasificación</option>
-          <option value="Type">Por tipos</option>
+          <option value="clasify">Clasificación por ...</option>
           <option value="Evolution">Por evolución</option>
           <option value="Legendary">Por Pokémon Legendario</option>
+        </select>
+
+        <select name="typePokemon" id="type">
+          <option value="types">Types....</option>
+          <option value="bug" id="bug">Bug</option>
+          <option value="dragon" id="dragon">Dragon</option>
+          <option value="fairy" id="fairy">Fairy</option>
+          <option value="fire" id="fire">Fire</option>
+          <option value="ghost" id="ghost">Ghost</option>
+          <option value="Ground" id="ground">Ground</option>
+          <option value="Psychic" id="psychic">Psychic</option>
+          <option value="Steel" id="steel">Steel</option>
+          <option value="dark" id="dark">Dark</option>
+          <option value="electric" id="electric">Electric</option>
+          <option value="fighting" id="fighting">Fighting</option>
+          <option value="flying" id="flying">Flying</option>
+          <option value="grass" id="grass">Grass</option>
+          <option value="ice" id="ice">Ice</option>
+          <option value="poison" id="poison">Poison</option>
+          <option value="water" id="water">Water</option>
+          <option value="rock" id="rock">Rock</option>
+          <option value="normal" id="normal">Normal</option>
         </select>
       </section>
       <section id="showData"></section>
