+++ src/style.css
@@ -122,7 +122,7 @@ h1 {
 
 #orderBy {
   margin-right: 4%;
-  margin-left: 26%;
+  margin-left: 8%;
   padding: 1%;
   border-radius: 10%;
   background-color: #f1b418;
@@ -134,11 +134,19 @@ h1 {
   border-radius: 10%;
   background-color: #f1b418;
   width: 20%;
-  margin-right: 26%;
+  margin-right: 4%;
   margin-left: 4%;
   height: auto;
 }
-
+#type {
+  padding: 1%;
+  border-radius: 10%;
+  background-color: #f1b418;
+  width: 20%;
+  margin-right: 8%;
+  margin-left: 4%;
+  height: auto;
+}
 #showData {
   display: flex;
   background-color: rgba(70, 123, 157, 0.7);
@@ -163,6 +171,7 @@ h1 {
 }
 
 #showData a {
+  text-decoration: none;
   width: 16%;
   margin: 1%;
 }
